Generalize method to TestHost
